Evolving a heteroclinic network
artificial neural network are often used to solve complex tasks in the eld of evolutionary robotic, but currently we don't know how the system can
learn by itself due to the high dimension and nonlinearity of the mathematical model behind it. Heteroclinic networks share many similar properties
with finite state machine, but are not discrete. Since the heteroclinic network is easy to interpret and model, it's interesting to analyze the behavior of the
system when we add a small-amplitude time-periodic forcing to the heteroclinic network. We use an evolutionary algorithm to evolve the system and try
to distinguish different added forcing.